"Changed list length variable type to long
The variable for listLength was changed from int to long in the methods that use it in the JavaSortsMain, JavaFXSortsMain, and SortingAlgorithm classes, and the appropriate changes were made to any methods that accessed it. This better accommodates large sorting quantities and avoids potential overflows."

+++ src/main/java/eu/lilithmonodia/javasorts/JavaFXSortsMain.java
@@ -79,7 +79,7 @@ public class JavaFXSortsMain extends Application {
                 return;
             }
 
-            int listLength;
+            long listLength;
             try {
                 listLength = Integer.parseInt(input);
             } catch (NumberFormatException exception) {
